requirements.txt
groq
pandas
sentence-transformers
faiss-cpu
python-dotenv